Update README.md for correct model pull
When I was setting up twinny today I followed the instructions here but the requests to the server would end up in 404. This was because as per docs the codelama model itself was pulled but not the instruct or code variants. The ollama OpenAI API will 404 and return {"error":"model 'XXX' not found, try pulling it first"}.